winkernel: add dualstack module
